3. validates SQLite entries before filtering and limiting branch results
